+++ team_task_manager/src/main/java/com/teamtask/team_task_manager/controller/ProjectController.java
@@ -1,8 +1,13 @@
 package com.teamtask.team_task_manager.controller;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
@@ -17,10 +22,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.teamtask.team_task_manager.model.Goal;
+import com.teamtask.team_task_manager.model.Membership;
 import com.teamtask.team_task_manager.model.Project;
+import com.teamtask.team_task_manager.model.Role;
 import com.teamtask.team_task_manager.model.Task;
 import com.teamtask.team_task_manager.model.User;
+import com.teamtask.team_task_manager.repository.EvaluationRepository;
 import com.teamtask.team_task_manager.repository.GoalRepository;
+import com.teamtask.team_task_manager.repository.MembershipRepository;
 import com.teamtask.team_task_manager.repository.ProjectRepository;
 import com.teamtask.team_task_manager.repository.TaskRepository;
 import com.teamtask.team_task_manager.repository.UserRepository;
@@ -37,18 +46,24 @@ public class ProjectController {
     private final GoalRepository goalRepository;
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
+    private final MembershipRepository membershipRepository;
+    private final EvaluationRepository evaluationRepository;
 
     @Autowired ProjectService projectService;
 
     public ProjectController(ProjectRepository projectRepository,
         GoalRepository goalRepository,
         TaskRepository taskRepository,
-        UserRepository userRepository
+        UserRepository userRepository,
+        MembershipRepository membershipRepository,
+        EvaluationRepository evaluationRepository
     ){
         this.projectRepository = projectRepository;
         this.goalRepository = goalRepository;
         this.taskRepository = taskRepository;
         this.userRepository = userRepository;
+        this.membershipRepository = membershipRepository;
+        this.evaluationRepository = evaluationRepository;
     }
 
     // Отображение списка проектов
@@ -70,6 +85,10 @@ public class ProjectController {
         model.addAttribute("tasksByGoal", projectService.GetTasksByGoal(id));
         model.addAttribute("tasksWithoutGoal", taskRepository.findByProjectIdAndGoalIsNull(id));
 
+        List<User> members = membershipRepository.findAllByProjectId(id).stream()
+            .map(Membership::getUser).collect(Collectors.toList());
+        model.addAttribute("members", members);
+
         model.addAttribute("isKanban", false);
         
         return "task-list";
@@ -85,9 +104,62 @@ public class ProjectController {
 
         model.addAttribute("tasksByStatus", projectService.GetTasksByStatus(id));
         
+        List<User> members = membershipRepository.findAllByProjectId(id).stream()
+            .map(Membership::getUser).collect(Collectors.toList());
+        model.addAttribute("members", members);
+
         model.addAttribute("isKanban", true);
 
         return "kanban";
+    }
+
+    // Просмотр личных задач проекта
+    @GetMapping("/{id}/my-tasks")
+    public String ShowProjectMyTasks(@PathVariable Long id, Model model){
+        if (!projectService.IsCurrentExecutor(id)){
+            throw new AccessDeniedException("Только у исполнителя есть личные задачи");
+        }
+
+        Project project = projectService.GetProjectIfAccesible(id);
+            
+        model.addAttribute("currentProject", project);
+        model.addAttribute("projects", projectService.GetUserProjects());
+
+        model.addAttribute("myTasks", projectService.GetPersonalTasksByProject(id));
+        
+        List<User> members = membershipRepository.findAllByProjectId(id).stream()
+            .map(Membership::getUser).collect(Collectors.toList());
+        model.addAttribute("members", members);
+
+        model.addAttribute("isKanban", false);
+
+        return "my-tasks";
+    }
+
+    // Переход на страницу оценки задач проекта
+    @GetMapping("/{id}/review")
+    public String ShowProjectReview(@PathVariable Long id, Model model){
+        if (!projectService.IsCurrentTester(id)){
+            throw new AccessDeniedException("Только тестировщик может оценивать задачи");
+        }
+
+        Project project = projectService.GetProjectIfAccesible(id);
+
+        model.addAttribute("currentProject", project);
+        model.addAttribute("projects", projectService.GetUserProjects());
+
+        model.addAttribute("reviewTasks", projectService.GetReviewTasksWithoutEvaluation(id));
+
+        User curUser = projectService.GetCurrentUser();
+        model.addAttribute("myEvaluations", evaluationRepository.findByReviewerId(curUser.getId()));
+        
+        List<User> members = membershipRepository.findAllByProjectId(id).stream()
+            .map(Membership::getUser).collect(Collectors.toList());
+        model.addAttribute("members", members);
+
+        model.addAttribute("isKanban", false);
+
+        return "review";
     }
 
     // Показать форму добавления проекта
@@ -106,8 +178,14 @@ public class ProjectController {
             return "project-form";
 
         User currentUser = projectService.GetCurrentUser();
-        project.setOwner(currentUser);
-        project.getMembers().add(currentUser);
+
+        Set<Role> roles = new HashSet<>(Arrays.asList(Role.values()));
+
+        // Тут мы добавляем лидера в члены проекта
+        Membership membership = new Membership(null, 
+            project, currentUser, roles);
+
+        project.getMembers().add(membership);
 
         Project saved = projectRepository.save(project);
         redirectAttributes.addFlashAttribute("message", "Проект успешно добавлен!");        
@@ -119,6 +197,10 @@ public class ProjectController {
     public String ShowEditForm(@PathVariable Long id, Model model){
         if (!projectService.HasAccess(id)){
             throw new AccessDeniedException("У вас нет доступа к этому проекту");
+        }
+
+        if (!projectService.IsCurrentLeader(id)){
+            throw new AccessDeniedException("Только лидер проекта может изменять проект");
         }
 
         Project project = projectRepository.findById(id)
@@ -135,6 +217,10 @@ public class ProjectController {
             RedirectAttributes redirectAttributes) {
         if (!projectService.HasAccess(id)){
             throw new AccessDeniedException("У вас нет доступа к этому проекту");
+        }
+
+        if (!projectService.IsCurrentLeader(id)){
+            throw new AccessDeniedException("Только лидер проекта может изменять проект");
         }
 
         if (result.hasErrors())
@@ -155,6 +241,10 @@ public class ProjectController {
             throw new AccessDeniedException("У вас нет доступа к этому проекту");
         }
             
+        if (!projectService.IsCurrentLeader(id)){
+            throw new AccessDeniedException("Только лидер проекта может удалять проект");
+        }
+
         projectRepository.deleteById(id);
         redirectAttributes.addFlashAttribute("message", "Проект успешно удален!");
         return "redirect:/projects";
@@ -167,9 +257,16 @@ public class ProjectController {
             throw new AccessDeniedException("У вас нет доступа к этому проекту");
         }
 
+        if (!projectService.IsCurrentLeader(id) && !projectService.IsCurrentManager(id)){
+            throw new AccessDeniedException("Только лидер или менеджер проекта может приглашать в проект");
+        }
+
         Project project = projectRepository.findById(id)
             .orElseThrow(() -> new IllegalArgumentException("Invalid project id: " + id));
         model.addAttribute("project", project);
+
+        Role[] availableRoles = {Role.EXECUTOR, Role.MANAGER, Role.TESTER};
+        model.addAttribute("availableRoles", availableRoles);
 
         return "invite-form";
     }
@@ -177,33 +274,145 @@ public class ProjectController {
     @PostMapping("/{id}/invite")
     public String InviteUser(@PathVariable Long id,
             @RequestParam String username,
+            @RequestParam(required = false) List<String> checkedRoles,
             RedirectAttributes redirectAttributes) {
         if (!projectService.HasAccess(id)){
             throw new AccessDeniedException("У вас нет доступа к этому проекту");
         }
 
-        Project currentProject = projectRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Project not found"));
-            
-        User currentUser = projectService.GetCurrentUser();
-
-        if (!currentProject.getOwner().getId().equals(currentUser.getId())){
-            redirectAttributes.addFlashAttribute("error", "Только владелец проекта может добавлять участников");
+        if (!projectService.IsCurrentLeader(id) && !projectService.IsCurrentManager(id)){
+            redirectAttributes.addFlashAttribute("error", "Только лидер или менеджер проекта может приглашать участников в проект");
             return "redirect:/projects/" + id;
         }
 
-        User newMember = userRepository.findByUsername(username)
-            .orElseThrow(() -> new RuntimeException("User not found"));
+        Project currentProject = projectRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Project not found"));
 
-        if (currentProject.getMembers().contains(newMember)){
+        User user = userRepository.findByUsername(username)
+            .orElseThrow(() -> new RuntimeException("User not found"));
+            
+        if (membershipRepository.findByProjectIdAndUserId(id, user.getId()).isPresent()){
             redirectAttributes.addFlashAttribute("error", "Участник уже есть в проекте");
             return "redirect:/projects/" + id;
         }
 
-        currentProject.getMembers().add(newMember);
-        projectRepository.save(currentProject);
+        Set<Role> roles = new HashSet<>();
+        if (checkedRoles != null){
+            roles = checkedRoles.stream()
+            .map(Role::valueOf)
+            .collect(Collectors.toSet());
+        }
+            
+        Membership newMember = new Membership(null, currentProject, user, roles);
 
-        redirectAttributes.addFlashAttribute("message", "Пользователь" + newMember.getUsername() + "добавлен в проект!");
+        membershipRepository.save(newMember);
+        // currentProject.getMembers().add(newMember);
+        // projectRepository.save(currentProject);
+
+        redirectAttributes.addFlashAttribute("message", "Пользователь " + newMember.getUser().getUsername() + " добавлен в проект!");
+        return "redirect:/projects/" + id;
+    }
+
+    // Исключение участника
+    @GetMapping("/{id}/exclude")
+    public String ExcludeMember(
+            @PathVariable Long id,
+            @RequestParam String username,
+            RedirectAttributes redirectAttributes){
+        if (!projectService.HasAccess(id)){
+            throw new AccessDeniedException("У вас нет доступа к этому проекту");
+        }
+        
+        if (!projectService.IsCurrentLeader(id) && !projectService.IsCurrentManager(id)){
+            throw new AccessDeniedException("Только лидер или менеджер проекта может исключать участника из проекта");
+        }
+
+        User excludeUser = userRepository.findByUsername(username)
+            .orElseThrow(() -> new RuntimeException("User not found"));
+        
+        Membership membership = membershipRepository.findByProjectIdAndUserId(id, excludeUser.getId())
+            .orElseThrow(() -> new RuntimeException("Membership not found"));
+        
+        if (membership.getRoles().contains(Role.LEADER)){
+            throw new AccessDeniedException("Лидер проекта не может исключить себя");
+        }
+
+        Project currentProject = projectRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Project not found"));
+        
+        currentProject.getMembers().remove(membership);
+        membershipRepository.deleteById(membership.getId());
+        return "redirect:/projects/" + id;
+    }
+
+    // Показать форму изменения ролей
+    @GetMapping("/{id}/editRoles")
+    public String ShowEditRolesForm(@PathVariable Long id, 
+            @RequestParam String username,
+            Model model){
+        if (!projectService.HasAccess(id)){
+            throw new AccessDeniedException("У вас нет доступа к этому проекту");
+        }
+
+        if (!projectService.IsCurrentLeader(id)){
+            throw new AccessDeniedException("Только лидер проекта может изменять роли участников проекта");
+        }
+
+        User editUser = userRepository.findByUsername(username)
+            .orElseThrow(() -> new RuntimeException("User not found"));
+        
+        Membership membership = membershipRepository.findByProjectIdAndUserId(id, editUser.getId())
+            .orElseThrow(() -> new RuntimeException("Membership not found"));
+        
+        if (membership.getRoles().contains(Role.LEADER)){
+            throw new AccessDeniedException("Лидер проекта не может изменить свои роли");
+        }
+
+        Project project = projectRepository.findById(id)
+            .orElseThrow(() -> new IllegalArgumentException("Invalid project id: " + id));
+        model.addAttribute("project", project);
+
+        Role[] availableRoles = {Role.EXECUTOR, Role.MANAGER, Role.TESTER};
+
+        model.addAttribute("availableRoles", availableRoles);
+        model.addAttribute("username", username);
+        model.addAttribute("checkedRoles", membership.getRoles());
+
+        return "invite-form";
+    }
+
+    // Обновление ролей участника проекта
+    @PostMapping("/{id}/editRoles")
+    public String UpdateRoles(@PathVariable Long id,
+            @RequestParam String username,
+            @RequestParam(required = false) List<String> checkedRoles,
+            RedirectAttributes redirectAttributes) {
+        if (!projectService.HasAccess(id)){
+            throw new AccessDeniedException("У вас нет доступа к этому проекту");
+        }
+            
+        if (!projectService.IsCurrentLeader(id)){
+            redirectAttributes.addFlashAttribute("error", "Только лидер проекта может изменять роли участников проекта");
+            return "redirect:/projects/" + id;
+        }
+
+        User user = userRepository.findByUsername(username)
+            .orElseThrow(() -> new RuntimeException("User not found"));
+        
+        Set<Role> roles = new HashSet<>();
+        if (checkedRoles != null){
+            roles = checkedRoles.stream()
+            .map(Role::valueOf)
+            .collect(Collectors.toSet());
+        }
+            
+        Membership editMember = membershipRepository.findByProjectIdAndUserId(id, user.getId())
+            .orElseThrow(() -> new RuntimeException("Membership not found"));
+
+        editMember.setRoles(roles);
+
+        membershipRepository.save(editMember);
+
         return "redirect:/projects/" + id;
     }
 }
